fastapi_template/requirements.txt
# FastAPI 및 웹 서버
fastapi
uvicorn
pydantic
pydantic-settings
email-validator

# 데이터베이스
sqlalchemy
alembic
psycopg2-binary
asyncpg

# 인증 및 보안
PyJWT
passlib
bcrypt
PyNaCl

# 캐싱 및 세션
redis

# 유틸리티
python-dotenv
httpx
tenacity
python-multipart


# 모듈
## validators
python-magic-bin
Pillow

# 개발 및 테스트 도구
mypy
pytest
pytest-asyncio
types-passlib
pytest-cov